requirements.txt
requests
beautifulsoup4
pandas
python-dateutil
matplotlib
numpy
Pillow
